test-a2a/pyproject.toml
[project]
name = "helloworld"
version = "0.1.0"
description = "HelloWorld agent example that only returns Messages"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk>=0.3.0",
    "click>=8.1.8",
    "httpx>=0.28.1",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "sse-starlette>=2.3.5",
    "starlette>=0.46.2",
    "uvicorn>=0.34.2",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
